+++ Dyalect/Runtime/Types/DyTuple.cs
@@ -252,13 +252,13 @@ namespace Dyalect.Runtime.Types
                 case "snd":
                     return DyForeignFunction.Member(name, GetSecond, -1, Statics.EmptyParameters);
                 default:
-                    return DyForeignFunction.Static("$$$AutoInvoke", (c, self) =>
+                    return DyForeignFunction.Auto(AutoKind.Generated, (c, self) =>
                     {
                         var idx = self.GetOrdinal(name);
                         if (idx == -1)
-                            return ctx.OperationNotSupported(name, self);
+                            return ctx.IndexOutOfRange(DyTypeNames.Tuple, name);
                         return self.GetItem(idx, ctx);
-                    }, -1, new Par("self"));
+                    });
             }
         }