+++ Docs/Changes.md
@@ -1,5 +1,5 @@
 # 0.7.6
-  * Done: _Add Tuple.new static method ([Issue #153](https://github.com/vorov2/dyalect/issues/153), [docs](https://github.com/vorov2/dyalect/wiki/Tuple#new)).
+  * Done: _Add `Tuple.new` static method_ ([Issue #153](https://github.com/vorov2/dyalect/issues/153), [docs](https://github.com/vorov2/dyalect/wiki/Tuple#new)).
   * Done: _`obj.Has(...)` optimization for simple cases_ ([Issue #156](https://github.com/vorov2/dyalect/issues/156)).
   * Access to tuple elements using member access syntax (e.g. `tuple.fieldName`) is refactored and redesigned for better efficiency.
   * Fixed: _Regression for member check (`?`) operator_ ([Issue #157](https://github.com/vorov2/dyalect/issues/157)).